converting most things to doubles, and using rounding for interpolation to the pixel to draw at
-changed most int vars regarding position and size to doubles
-added roundAndCast() method to primitive, which rounds the number to
the nearest 1 and casts to integer. makes sending draw calls less
tedious
-removed vestigial textbox class, which was a remainder from the
separation from stardust game

+++ ClosedPoly.java
@@ -179,8 +179,8 @@ public class ClosedPoly extends Primitive {
 		updatePoints();
 	}
 
-	public int[] getpos() {
-        return new int[]{ origin[0], origin[1] };
+	public double[] getpos() {
+        return new double[]{ origin[0], origin[1] };
 	}
 
 }
